requirements.txt
toposort
pywin32
mouse==0.7.1
MouseInfo==0.1.3
numpy==1.19.5
opencv-python==4.5.5.62
Pillow==8.4.0
PyAutoGUI==0.9.53
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.5
pyperclip==1.8.2
pyrealsense2==2.50.0.3812
PyRect==0.1.4
PyScreeze==0.1.28
pytweening==1.0.4
pywin32==303
scipy==1.5.4
six==1.16.0